scx_bpfland: handle per-CPU tasks preempted by a higher scheduling class
When a CPU is stolen by a higher scheduling class we re-schedule all the
tasks in the local DSQ from ops.cpu_release() to give them a chance to
be dispatched on a different CPU.

However, this is pointless for per-CPU tasks, since they can only run on
that specific CPU, so re-scheduling them is only increasing the risk of
triggering stalls.

Therefore, dispatch them directly on reschedule, to maximize their
chances of reclaiming the CPU quickly and avoiding stalls.

This approach will be more effective once dl_server support is added to
the sched_ext core.

Signed-off-by: Andrea Righi <arighi@nvidia.com>